+++ src/shared/utils/Logging/generateErrorBasedOnEnvironment.spec.ts
