+++ src/shared/utils/Logging/generateErrorBasedOnEnvironment.ts
